moved runprogram to separate package. preparing for 2d shapes program.
