+++ VolumesSurfaceArea/nbproject/project.properties
@@ -54,7 +54,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=volumessurfacearea.RunProgram
+main.class=excecuteProgram.RunProgram
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
